requirements.txt
azure-functions>=1.20.0
requests
flask
gunicorn
